# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión E-commerce",
        "description": "API para gestión de órdenes de venta de un e-commerce",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener órdenes del usuario autenticado con filtros y paginación",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "processing",
                                "shipped",
                                "delivered",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "filter[min_total]",
                        "in": "query",
                        "description": "Total mínimo",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "filter[max_total]",
                        "in": "query",
                        "description": "Total máximo",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "filter[date_from]",
                        "in": "query",
                        "description": "Fecha desde",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "filter[date_to]",
                        "in": "query",
                        "description": "Fecha hasta",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenar por campo (ej: -created_at, total, status)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página (máximo 50)",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Order"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear una nueva orden",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "address_id",
                                    "products"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "address_id": {
                                        "type": "integer"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener el detalle de una orden",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orden no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/stats": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener estadísticas de órdenes",
                "operationId": "5c3f86e9e2d05f0dbc3fc8bcf13fc393",
                "responses": {
                    "200": {
                        "description": "Estadísticas de órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_orders": {
                                            "type": "integer"
                                        },
                                        "pending_orders": {
                                            "type": "integer"
                                        },
                                        "completed_orders": {
                                            "type": "integer"
                                        },
                                        "total_revenue": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProductDetail": {
                "title": "Product Detail",
                "description": "Modelo detallado de producto del e-commerce",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del producto",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "maxLength": 255,
                        "example": "iPhone 15 Pro"
                    },
                    "description": {
                        "description": "Descripción detallada del producto",
                        "type": "string",
                        "example": "Último modelo de iPhone con chip A17 Pro"
                    },
                    "price": {
                        "description": "Precio del producto en USD",
                        "type": "number",
                        "format": "float",
                        "minimum": 0.01,
                        "example": 1199.99
                    },
                    "stock": {
                        "description": "Cantidad disponible en inventario",
                        "type": "integer",
                        "minimum": 0,
                        "example": 50
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:30:45.000000Z"
                    },
                    "deleted_at": {
                        "description": "Fecha y hora de eliminación lógica (null si está activo)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Address": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "street": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "zip": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Invoice": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "order_id": {
                        "type": "integer"
                    },
                    "billing_data": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "address_id": {
                        "type": "integer"
                    },
                    "invoice_id": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductDetail"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "invoice": {
                        "$ref": "#/components/schemas/Invoice"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}